+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Upgrade Vert.x to 4.3.5
 * Moved from using the Jaeger exporter to OTLP exporter by default
 * Kafka Exporter support for `Recreate` deployment strategy
+* `ImageStream` validation for Kafka Connect builds on OpenShift
 
 ### Changes, deprecations and removals